Generate correct bindings for constructor functions
