+++ cs-bindgen-cli/src/generate/func.rs
@@ -4,51 +4,6 @@ use proc_macro2::TokenStream;
 use quote::*;
 use syn::{punctuated::Punctuated, token::Comma, Ident};
 
-pub fn quote_raw_binding(bindgen_fn: &BindgenFn, dll_name: &str) -> TokenStream {
-    let entry_point = bindgen_fn.generated_name();
-    let binding_return_ty = match bindgen_fn.ret.primitive() {
-        None => quote! { void },
-        Some(prim) => match prim {
-            Primitive::String => quote! { RustOwnedString },
-            Primitive::Char => quote! { uint },
-            Primitive::I8 => quote! { sbyte },
-            Primitive::I16 => quote! { short },
-            Primitive::I32 => quote! { int },
-            Primitive::I64 => quote! { long },
-            Primitive::U8 => quote! { byte },
-            Primitive::U16 => quote! { ushort },
-            Primitive::U32 => quote! { uint },
-            Primitive::U64 => quote! { ulong },
-            Primitive::F32 => quote! { float },
-            Primitive::F64 => quote! { double },
-            Primitive::Bool => quote! { byte },
-        },
-    };
-
-    let mut binding_args = bindgen_fn
-        .args
-        .iter()
-        .map(|arg| {
-            let ident = arg.ident();
-            let ty = quote_primitive_binding_arg(arg.ty);
-            quote! { #ty #ident }
-        })
-        .collect::<Punctuated<_, Comma>>();
-
-    if bindgen_fn.receiver.is_some() {
-        binding_args.insert(0, quote! { *void self })
-    }
-
-    let raw_ident = bindgen_fn.generated_ident();
-    quote! {
-        [DllImport(
-            #dll_name,
-            EntryPoint = #entry_point,
-            CallingConvention = CallingConvention.Cdecl)]
-        internal static extern #binding_return_ty #raw_ident(#binding_args);
-    }
-}
-
 pub fn quote_wrapper_body(bindgen_fn: &BindgenFn, output: &Ident) -> TokenStream {
     // Build the list of arguments to the wrapper function.
     let mut invoke_args = bindgen_fn
